feature(cl): nav for accounts, fix paths, add app_name for aps
