+++ accounts/templates/order_details.html
@@ -2,7 +2,9 @@
 {% load static %}
 {% block content %}
 <div>
+    <p><a href="{% url 'accounts:profile' %}">Profile</a></p>
     <h3>Order Details</h3>
+    <a href="{% url 'accounts:orders' %}">Back to All Orders</a>
     <div>
         <p>{{ viewed_order.customer }}</p>
         <p>Complete: {{ viewed_order.complete }}</p>
