+++ accounts/templates/order_history.html
@@ -3,14 +3,14 @@
 {% block content %}
 
 <div>
+    <p><a href="{% url 'accounts:profile' %}">Profile</a></p>
+    <p><a href="{% url 'store:favourites' %}">Favourites</a></p>
     <p>SIMPLE NAV JUST FOR NOW</p>
-    <p><a href="{% url 'orders_history' %}">Orders History</a></p>
-    <p><a href="{% url 'favourites' %}">Favourites</a></p>
     <p>Products in Cart: {{ cart_items }}</p>
 
     <div>
         {% for category in categories %}
-            <strong><a href="{% url 'category' category.id %}">{{ category.name }}</a></strong>
+            <strong><a href="{% url 'store:category' category.id %}">{{ category.name }}</a></strong>
         {% endfor %}
     </div>
 </div>
@@ -18,7 +18,7 @@
 <div>
     <h3>Order History</h3>
     <hr>
-    <a href="{% url 'orders_history' %}">Back to Orders History</a>
+    <a href="{% url 'accounts:orders_history' %}">Back to Orders History</a>
     <div>
         {{ customer }}
         <hr>
