+++ accounts/templates/orders.html
@@ -3,8 +3,9 @@
 {% block content %}
 
 <div>
+    <p><a href="{% url 'accounts:profile' %}">Profile</a></p>
     <h1>All Orders</h1>
-    <a href="{% url 'orders_completed' %}">Orders Completed</a>
+    <a href="{% url 'accounts:orders_completed' %}">Orders Completed</a>
     <ul>
         {% for order in object_list %}
             <ul>
@@ -12,7 +13,7 @@
                 <p>Customer: {{ order.customer }} - Complete: {{ order.complete }}</p>
                 <p>Date: {{ order.date_ordered.date }} - Time: {{ order.date_ordered.time }}</p>
                 <p>Total: {{ order.get_cart_total }} - Items: {{ order.get_cart_items }}</p>
-                <a href="{% url 'order_details' order.id %}">Order details</a>
+                <a href="{% url 'accounts:order_details' order.id %}">Order details</a>
                 <hr>
             </ul>
         {% endfor %}
