+++ accounts/templates/profile.html
@@ -2,7 +2,23 @@
 {% load static %}
 {% block content %}
     <div>
-        <p>Hello, {{ customer.name }}!</p>
+        <p><a href="{% url 'accounts:login' %}">Login</a></p>
+        <p><a href="{% url 'accounts:logout' %}">Logout</a></p>
+        <p><a href="{% url 'accounts:password_change' %}">Change password</a></p>
+        <p><a href="{% url 'accounts:sign_up' %}">Sign Up</a></p>
+        <p><a href="{% url 'accounts:profile' %}">Profile</a></p>
+        <p><a href="{% url 'accounts:orders_history' %}">Orders History</a></p>
+        <p><a href="{% url 'accounts:orders' %}">Orders</a></p>
+
+
+
+
+
+
+
+    </div>
+    <div>
+        <p>Hello, {{ user }}!</p>
         {% for order in object_list %}
             {{ order.id }}
         {% endfor %}
