+++ accounts/views.py
@@ -1,11 +1,18 @@
+from django.contrib.admin.views.decorators import staff_member_required
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.auth.views import LoginView, PasswordChangeView
+from django.shortcuts import render
 from django.urls import reverse_lazy
 from django.views.generic import CreateView, ListView
-
+from django.contrib.auth.mixins import PermissionRequiredMixin, UserPassesTestMixin
 from store.models import Customer, Order, OrderItem
 
 from .forms import SignUpForm
+
+
+class StaffRequiredMixin(UserPassesTestMixin):
+    def test_func(self):
+        return self.request.user.is_staff
 
 
 class SubmittableLoginView(LoginView):
@@ -38,4 +45,60 @@ class ProfileView(LoginRequiredMixin, ListView):
     #         'orders_items': history_order.get_orderitems,
     #     })
     #     return context
+
+
+# view for Customer
+def orders_history(request):
+    if request.user.is_authenticated:
+        customer = request.user.customer
+        user_orders = Order.objects.filter(customer=customer, complete=True)
+        context = {'customer': customer, 'user_orders': user_orders}
+        return render(request, 'orders_history.html', context)
+
+
+# view for Customer
+def order_history(request, user_order_id):
+    if request.user.is_authenticated:
+        customer = request.user.customer
+        history_order = Order.objects.get(customer=customer, id=user_order_id)
+        history_items = history_order.get_orderitems
+
+        context = {'history_order': history_order, 'history_items': history_items}
+        return render(request, 'order_history.html', context)
+
+
+# view for Admin
+class OrdersView(StaffRequiredMixin, PermissionRequiredMixin, ListView):
+    template_name = 'orders.html'
+    model = Order
+    permission_required = 'store/.orders'
+
+
+# view for Admin
+@staff_member_required
+def order_details(request, order_details_id):
+    viewed_order = Order.objects.get(id=order_details_id)
+    order_items = viewed_order.get_orderitems
+    order_comment = viewed_order.comment
+    context = {'viewed_order': viewed_order, 'order_items': order_items, 'order_comment': order_comment}
+    return render(request, 'order_details.html', context)
+
+
+# view for Admin
+class OrdersCompletedView(StaffRequiredMixin, PermissionRequiredMixin, ListView):
+    template_name = 'orders_completed.html'
+    model = Order
+    permission_required = 'accounts/.orders_completed'
+
+    def get_queryset(self):
+        return Order.objects.filter(complete=True)
+
+
+# view for Admin
+@staff_member_required
+def completed_order_details(request, completed_order_details_id):
+    viewed_order_completed = Order.objects.get(id=completed_order_details_id)
+    order_items = viewed_order_completed.get_orderitems
+    context = {'viewed_order_completed': viewed_order_completed, 'order_items': order_items}
+    return render(request, 'completed_order_details.html', context)
