+++ store/views.py
@@ -1,17 +1,11 @@
 import datetime
 
 
-from django.contrib.auth.decorators import login_required
-from django.contrib.admin.views.decorators import staff_member_required
 from django.shortcuts import render, HttpResponseRedirect, redirect, get_object_or_404
 from django.http import JsonResponse, HttpResponse
 from django.urls import reverse, reverse_lazy
 from django.contrib import messages
-from django.views.generic import CreateView, ListView, DetailView
-from django.contrib.auth.mixins import PermissionRequiredMixin, UserPassesTestMixin
-from django.views.generic import ListView, CreateView
-from django.contrib.auth.mixins import PermissionRequiredMixin, UserPassesTestMixin
-#from django.contrib.admin.views.decorators.staff_member_required import
+from django.views.generic import ListView
 
 import json
 
@@ -22,11 +16,6 @@ from .forms import ProductOpinionForm, OrderForm
 from .utils import *
 
 
-class StaffRequiredMixin(UserPassesTestMixin):
-    def test_func(self):
-        return self.request.user.is_staff
-
-    
 def set_initial_cart_status(request):
     if request.user.is_authenticated:
         customer = request.user.customer
@@ -39,38 +28,6 @@ def set_initial_cart_status(request):
         cart_items = order['get_cart_items']
     context = {'cart_items': cart_items}
     return context
-
-
-class OrdersView(StaffRequiredMixin, PermissionRequiredMixin, ListView):
-    template_name = 'orders.html'
-    model = Order
-    permission_required = 'store/.orders'
-
-
-@staff_member_required
-def order_details(request, order_details_id):
-    viewed_order = Order.objects.get(id=order_details_id)
-    order_items = viewed_order.get_orderitems
-    order_comment = viewed_order.comment
-    context = {'viewed_order': viewed_order, 'order_items': order_items, 'order_comment': order_comment}
-    return render(request, 'order_details.html', context)
-
-
-class OrdersCompletedView(StaffRequiredMixin, PermissionRequiredMixin, ListView):
-    template_name = 'orders_completed.html'
-    model = Order
-    permission_required = 'store/.orders_completed'
-
-    def get_queryset(self):
-        return Order.objects.filter(complete=True)
-
-
-@staff_member_required
-def completed_order_details(request, completed_order_details_id):
-    viewed_order_completed = Order.objects.get(id=completed_order_details_id)
-    order_items = viewed_order_completed.get_orderitems
-    context = {'viewed_order_completed': viewed_order_completed, 'order_items': order_items}
-    return render(request, 'completed_order_details.html', context)
 
 
 def home(request):
@@ -275,24 +232,6 @@ def meta_product(request, meta_product_id):
                'opinions': opinions,
                'cart_items': cart_items}
     return render(request, 'meta_product.html', context)
-
-
-def orders_history(request):
-    if request.user.is_authenticated:
-        customer = request.user.customer
-        user_orders = Order.objects.filter(customer=customer, complete=True)
-        context = {'customer': customer, 'user_orders': user_orders}
-        return render(request, 'orders_history.html', context)
-
-
-def order_history(request, user_order_id):
-    if request.user.is_authenticated:
-        customer = request.user.customer
-        history_order = Order.objects.get(customer=customer, id=user_order_id)
-        history_items = history_order.get_orderitems
-
-        context = {'history_order': history_order, 'history_items': history_items}
-        return render(request, 'order_history.html', context)
 
 
 def favourites(request):
